requirements.txt
Django==4.1.4
environs==9.5.0
pillow==9.3.0
django-phonenumber-field==7.0.1
phonenumbers==8.13.2